requirements.txt
discord.py==1.3.3
mcstatus